authoriztion added

+++ backend/routes/posts.js
@@ -34,8 +34,10 @@ multer({storage:storage}).single("image"),(req,res,next)=>{
     const post=new Post({
         title:req.body.title,
         content:req.body.content,
-        imagePath:url+"/images/"+ req.file.filename 
+        imagePath:url+"/images/"+ req.file.filename, 
+        creator:req.userData.userId
     });
+
     post.save().then(createdPost=>{
         res.status(201).json({
             message:'Post added Successfully',
@@ -99,20 +101,32 @@ multer({storage:storage}).single("image"),(req,res,next)=>{
         _id: req.body.id,
         title:req.body.title,
         content:req.body.content,
-        imagePath:imagePath
+        imagePath:imagePath,
+        creator:req.userData.userId
 
     })
-    Post.updateOne({_id: req.params.id},post).then(result=>{
-        res.status(200).json({message:'Update Successful!'})
+    Post.updateOne({_id: req.params.id,creator:req.userData.userId},post).then(result=>{
+        if(result.nModified>0){
+            res.status(200).json({message:'Update Successful!'})
+        }
+        else{
+            res.status(401).json({message:'Not authorized!'})
+        }
+       
     })
 });
 
 router.delete('/:id',
 checkauth,
 (req,res,next)=>{
-    Post.deleteOne({_id:req.params.id}).then(result =>{
-        console.log(result);
-        res.status(200).json({message:"Post Deleted!"})
+    Post.deleteOne({_id:req.params.id,creator:req.userData.userId}).then(result =>{
+        if(result.n>0){
+            res.status(200).json({message:"Post Deleted!"})
+        }
+        else{
+            res.status(401).json({message:'Not authorized!'})
+        }
+      
     });
 
 });
